pyproject.toml
[tool.poetry]
name = "banco-api-workshop"
version = "0.1.0"
description = ""
authors = ["Luiz Fernando <luizfsoliveira.lm@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11 || ^3.12"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.36"
psycopg2-binary = "^2.9.10"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
